% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{delete_byte_match_set}
\alias{delete_byte_match_set}
\title{Permanently deletes a ByteMatchSet}
\usage{
delete_byte_match_set(ByteMatchSetId, ChangeToken)
}
\arguments{
\item{ByteMatchSetId}{The \code{ByteMatchSetId} of the ByteMatchSet that you want to delete. \code{ByteMatchSetId} is returned by CreateByteMatchSet and by ListByteMatchSets.}

\item{ChangeToken}{The value returned by the most recent call to GetChangeToken.}
}
\description{
Permanently deletes a ByteMatchSet. You can't delete a \code{ByteMatchSet} if it's still used in any \code{Rules} or if it still includes any ByteMatchTuple objects (any filters).
}
\details{
If you just want to remove a \code{ByteMatchSet} from a \code{Rule}, use UpdateRule.

To permanently delete a \code{ByteMatchSet}, perform the following steps:
\enumerate{
\item Update the \code{ByteMatchSet} to remove filters, if any. For more information, see UpdateByteMatchSet.
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of a \code{DeleteByteMatchSet} request.
\item Submit a \code{DeleteByteMatchSet} request.
}
}
\examples{
# The following example deletes a byte match set with the ID
# exampleIDs3t-46da-4fdb-b8d5-abc321j569j5.
delete_byte_match_set(
  ByteMatchSetId = "exampleIDs3t-46da-4fdb-b8d5-abc321j569j5",
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f"
)

}
