% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{delete_regex_match_set}
\alias{delete_regex_match_set}
\title{Permanently deletes a RegexMatchSet}
\usage{
delete_regex_match_set(
  RegexMatchSetId = "string",
  ChangeToken = "string"
)
}
\arguments{
\item{RegexMatchSetId}{[required] The \code{RegexMatchSetId} of the RegexMatchSet that you want to delete. \code{RegexMatchSetId} is returned by CreateRegexMatchSet and by ListRegexMatchSets.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
Permanently deletes a RegexMatchSet. You can't delete a \code{RegexMatchSet} if it's still used in any \code{Rules} or if it still includes any \code{RegexMatchTuples} objects (any filters).
}
\details{
If you just want to remove a \code{RegexMatchSet} from a \code{Rule}, use UpdateRule.

To permanently delete a \code{RegexMatchSet}, perform the following steps:
\enumerate{
\item Update the \code{RegexMatchSet} to remove filters, if any. For more information, see UpdateRegexMatchSet.
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of a \code{DeleteRegexMatchSet} request.
\item Submit a \code{DeleteRegexMatchSet} request.
}
}
\examples{

}
