% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{delete_rule}
\alias{delete_rule}
\title{Permanently deletes a Rule}
\usage{
delete_rule(
  RuleId = "string",
  ChangeToken = "string"
)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the Rule that you want to delete. \code{RuleId} is returned by CreateRule and by ListRules.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
Permanently deletes a Rule. You can't delete a \code{Rule} if it's still used in any \code{WebACL} objects or if it still includes any predicates, such as \code{ByteMatchSet} objects.
}
\details{
If you just want to remove a \code{Rule} from a \code{WebACL}, use UpdateWebACL.

To permanently delete a \code{Rule} from AWS WAF, perform the following steps:
\enumerate{
\item Update the \code{Rule} to remove predicates, if any. For more information, see UpdateRule.
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of a \code{DeleteRule} request.
\item Submit a \code{DeleteRule} request.
}
}
\examples{
# The following example deletes a rule with the ID WAFRule-1-Example.
delete_rule(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f",
  RuleId = "WAFRule-1-Example"
)

}
