% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_geo_match_set}
\alias{get_geo_match_set}
\title{Returns the GeoMatchSet that is specified by \code{GeoMatchSetId}}
\usage{
get_geo_match_set(
  GeoMatchSetId = "string"
)
}
\arguments{
\item{GeoMatchSetId}{[required] The \code{GeoMatchSetId} of the GeoMatchSet that you want to get. \code{GeoMatchSetId} is returned by CreateGeoMatchSet and by ListGeoMatchSets.}
}
\description{
Returns the GeoMatchSet that is specified by \code{GeoMatchSetId}.
}
\examples{

}
