% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_permission_policy}
\alias{get_permission_policy}
\title{Returns the IAM policy attached to the RuleGroup}
\usage{
get_permission_policy(
  ResourceArn = "string"
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the RuleGroup for which you want to get the policy.}
}
\description{
Returns the IAM policy attached to the RuleGroup.
}
\examples{

}
