% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_rate_based_rule}
\alias{get_rate_based_rule}
\title{Returns the RateBasedRule that is specified by the RuleId that you included in the GetRateBasedRule request}
\usage{
get_rate_based_rule(RuleId)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the RateBasedRule that you want to get. \code{RuleId} is returned by CreateRateBasedRule and by ListRateBasedRules.}
}
\description{
Returns the RateBasedRule that is specified by the \code{RuleId} that you included in the \code{GetRateBasedRule} request.
}
\section{Accepted Parameters}{
\preformatted{get_rate_based_rule(
  RuleId = "string"
)
}
}

