% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_rule}
\alias{get_rule}
\title{Returns the Rule that is specified by the \code{RuleId} that you included in the \code{GetRule} request}
\usage{
get_rule(RuleId)
}
\arguments{
\item{RuleId}{The \code{RuleId} of the Rule that you want to get. \code{RuleId} is returned by CreateRule and by ListRules.}
}
\description{
Returns the Rule that is specified by the \code{RuleId} that you included in the \code{GetRule} request.
}
\examples{
# The following example returns the details of a rule with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
get_rule(
  RuleId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)

}
