% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_size_constraint_set}
\alias{get_size_constraint_set}
\title{Returns the SizeConstraintSet specified by \code{SizeConstraintSetId}}
\usage{
get_size_constraint_set(SizeConstraintSetId)
}
\arguments{
\item{SizeConstraintSetId}{The \code{SizeConstraintSetId} of the SizeConstraintSet that you want to get. \code{SizeConstraintSetId} is returned by CreateSizeConstraintSet and by ListSizeConstraintSets.}
}
\description{
Returns the SizeConstraintSet specified by \code{SizeConstraintSetId}.
}
\examples{
# The following example returns the details of a size constraint match set
# with the ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
get_size_constraint_set(
  SizeConstraintSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)

}
