% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_sql_injection_match_set}
\alias{get_sql_injection_match_set}
\title{Returns the SqlInjectionMatchSet that is specified by \code{SqlInjectionMatchSetId}}
\usage{
get_sql_injection_match_set(SqlInjectionMatchSetId)
}
\arguments{
\item{SqlInjectionMatchSetId}{The \code{SqlInjectionMatchSetId} of the SqlInjectionMatchSet that you want to get. \code{SqlInjectionMatchSetId} is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.}
}
\description{
Returns the SqlInjectionMatchSet that is specified by \code{SqlInjectionMatchSetId}.
}
\examples{
# The following example returns the details of a SQL injection match set
# with the ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
get_sql_injection_match_set(
  SqlInjectionMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)

}
