% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_web_acl}
\alias{get_web_acl}
\title{Returns the WebACL that is specified by \code{WebACLId}}
\usage{
get_web_acl(WebACLId)
}
\arguments{
\item{WebACLId}{The \code{WebACLId} of the WebACL that you want to get. \code{WebACLId} is returned by CreateWebACL and by ListWebACLs.}
}
\description{
Returns the WebACL that is specified by \code{WebACLId}.
}
\examples{
# The following example returns the details of a web ACL with the ID
# createwebacl-1472061481310.
get_web_acl(
  WebACLId = "createwebacl-1472061481310"
)

}
