% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{get_xss_match_set}
\alias{get_xss_match_set}
\title{Returns the XssMatchSet that is specified by \code{XssMatchSetId}}
\usage{
get_xss_match_set(XssMatchSetId)
}
\arguments{
\item{XssMatchSetId}{The \code{XssMatchSetId} of the XssMatchSet that you want to get. \code{XssMatchSetId} is returned by CreateXssMatchSet and by ListXssMatchSets.}
}
\description{
Returns the XssMatchSet that is specified by \code{XssMatchSetId}.
}
\examples{
# The following example returns the details of an XSS match set with the
# ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
get_xss_match_set(
  XssMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)

}
