% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{list_activated_rules_in_rule_group}
\alias{list_activated_rules_in_rule_group}
\title{Returns an array of ActivatedRule objects}
\usage{
list_activated_rules_in_rule_group(RuleGroupId = NULL,
  NextMarker = NULL, Limit = NULL)
}
\arguments{
\item{RuleGroupId}{The \code{RuleGroupId} of the RuleGroup for which you want to get a list of ActivatedRule objects.}

\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{ActivatedRules} than the value of \code{Limit}, AWS WAF returns a \code{NextMarker} value in the response that allows you to list another group of \code{ActivatedRules}. For the second and subsequent \code{ListActivatedRulesInRuleGroup} requests, specify the value of \code{NextMarker} from the previous response to get information about another batch of \code{ActivatedRules}.}

\item{Limit}{Specifies the number of \code{ActivatedRules} that you want AWS WAF to return for this request. If you have more \code{ActivatedRules} than the number that you specify for \code{Limit}, the response includes a \code{NextMarker} value that you can use to get another batch of \code{ActivatedRules}.}
}
\description{
Returns an array of ActivatedRule objects.
}
\section{Accepted Parameters}{
\preformatted{list_activated_rules_in_rule_group(
  RuleGroupId = "string",
  NextMarker = "string",
  Limit = 123
)
}
}

