% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_package.R
\docType{package}
\name{paws.waf-package}
\alias{paws.waf}
\alias{paws.waf-package}
\title{paws.waf: AWS WAF}
\description{
This is the *AWS WAF API Reference* for using AWS WAF with
Amazon CloudFront. The AWS WAF actions and data types listed in the
reference are available for protecting Amazon CloudFront
distributions. You can use these actions and data types via the
endpoint *waf.amazonaws.com*. This guide is for developers who need
detailed information about the AWS WAF API actions, data types, and
errors. For detailed information about AWS WAF features and an
overview of how to use the AWS WAF API, see the [AWS WAF Developer
Guide](http://docs.aws.amazon.com/waf/latest/developerguide/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
