% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.waf_operations.R
\name{update_geo_match_set}
\alias{update_geo_match_set}
\title{Inserts or deletes GeoMatchConstraint objects in an \code{GeoMatchSet}}
\usage{
update_geo_match_set(GeoMatchSetId, ChangeToken, Updates)
}
\arguments{
\item{GeoMatchSetId}{The \code{GeoMatchSetId} of the GeoMatchSet that you want to update. \code{GeoMatchSetId} is returned by CreateGeoMatchSet and by ListGeoMatchSets.}

\item{ChangeToken}{The value returned by the most recent call to GetChangeToken.}

\item{Updates}{An array of \code{GeoMatchSetUpdate} objects that you want to insert into or delete from an GeoMatchSet. For more information, see the applicable data types:
\itemize{
\item GeoMatchSetUpdate: Contains \code{Action} and \code{GeoMatchConstraint}
\item GeoMatchConstraint: Contains \code{Type} and \code{Value}

You can have only one \code{Type} and \code{Value} per \code{GeoMatchConstraint}. To add multiple countries, include multiple \code{GeoMatchSetUpdate} objects in your request.
}}
}
\description{
Inserts or deletes GeoMatchConstraint objects in an \code{GeoMatchSet}. For each \code{GeoMatchConstraint} object, you specify the following values:
}
\details{
\itemize{
\item Whether to insert or delete the object from the array. If you want to change an \code{GeoMatchConstraint} object, you delete the existing object and add a new one.
\item The \code{Type}. The only valid value for \code{Type} is \code{Country}.
\item The \code{Value}, which is a two character code for the country to add to the \code{GeoMatchConstraint} object. Valid codes are listed in GeoMatchConstraint\$Value.
}

To create and configure an \code{GeoMatchSet}, perform the following steps:
\enumerate{
\item Submit a CreateGeoMatchSet request.
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of an UpdateGeoMatchSet request.
\item Submit an \code{UpdateGeoMatchSet} request to specify the country that you want AWS WAF to watch for.
}

When you update an \code{GeoMatchSet}, you specify the country that you want to add and/or the country that you want to delete. If you want to change a country, you delete the existing country and add the new one.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the \href{http://docs.aws.amazon.com/waf/latest/developerguide/}{AWS WAF Developer Guide}.
}
\examples{

}
